// web/src/api/models/TypesNeighbor.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesNeighbor
 */
export interface TypesNeighbor {
    /**
     * 
     * @type {string}
     * @memberof TypesNeighbor
     */
    dev: string;
    /**
     * 
     * @type {string}
     * @memberof TypesNeighbor
     */
    dst: string;
    /**
     * 
     * @type {string}
     * @memberof TypesNeighbor
     */
    family: string;
    /**
     * 
     * @type {string}
     * @memberof TypesNeighbor
     */
    lladdr?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesNeighbor
     */
    state: string;
}

/**
 * Check if a given object implements the TypesNeighbor interface.
 */
export function instanceOfTypesNeighbor(value: object): value is TypesNeighbor {
    if (!('dev' in value) || value['dev'] === undefined) return false;
    if (!('dst' in value) || value['dst'] === undefined) return false;
    if (!('family' in value) || value['family'] === undefined) return false;
    if (!('state' in value) || value['state'] === undefined) return false;
    return true;
}

export function TypesNeighborFromJSON(json: any): TypesNeighbor {
    return TypesNeighborFromJSONTyped(json, false);
}

export function TypesNeighborFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesNeighbor {
    if (json == null) {
        return json;
    }
    return {
        
        'dev': json['dev'],
        'dst': json['dst'],
        'family': json['family'],
        'lladdr': json['lladdr'] == null ? undefined : json['lladdr'],
        'state': json['state'],
    };
}

export function TypesNeighborToJSON(json: any): TypesNeighbor {
    return TypesNeighborToJSONTyped(json, false);
}

export function TypesNeighborToJSONTyped(value?: TypesNeighbor | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'dev': value['dev'],
        'dst': value['dst'],
        'family': value['family'],
        'lladdr': value['lladdr'],
        'state': value['state'],
    };
}